create pages design revisions popup dialogs changed reminder database 'where' parameter changed

+++ ess_app/lib/guardian/memory/memory_diary_tab.dart
@@ -1,10 +1,12 @@
 import 'package:awesome_dialog/awesome_dialog.dart';
 import 'package:ess_app/guardian/edit/edit_entry_diary.dart';
+import 'package:ess_app/guardian/widgets/popup_dialogs.dart';
 import 'package:ess_app/models/diary_model.dart';
 import 'package:ess_app/services/database.dart';
 import 'package:ess_app/utils/colors.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:page_transition/page_transition.dart';
 
 import '../widgets/memory_tab_listview.dart';
 import 'package:ess_app/constants.dart';
@@ -45,14 +47,25 @@ class _MemoryDiaryTabState extends State<MemoryDiaryTab> {
                           final docs = data[index];
                           return MemoryTabListView(
                               diary: docs,
-                              deleteTapped: (context) =>
-                                  deleteDiaryEntry(diaryCollection, docs.uid),
+                              deleteTapped: (context) async {
+                                try {
+                                  bool? deleteConfirmed = await showConfirmationDialog(context, 'Are you sure you want to delete?');
+                                  if (deleteConfirmed == true) {
+                                    // perform deletion
+                                    await deleteDiaryEntry(diaryCollection, docs.uid);
+                                  } else {
+                                    // user canceled deletion
+                                  }
+                                } catch (e) {
+                                  // handle any errors that might occur here
+                                  print(e);
+                                }
+                              },
                               editTapped: (context) =>
                                   editDiaryEntry(context, docs));
                         });
                   }
-                } else if (snapshot.connectionState ==
-                    ConnectionState.waiting) {
+                } else if (snapshot.connectionState == ConnectionState.waiting) {
                   return Center(child: CircularProgressIndicator());
                 } else {
                   return Text('Error: $snapshot.error');
@@ -65,84 +78,19 @@ class _MemoryDiaryTabState extends State<MemoryDiaryTab> {
     );
   }
 
-  // delete? yes or no
-  AwesomeDialog deleteDialog(BuildContext context, String index) {
-    return AwesomeDialog(
-      context: context,
-      dialogType: DialogType.QUESTION,
-      borderSide: BorderSide(
-        color: AppColors.secondColor,
-        width: 2,
-      ),
-      width: MediaQuery.of(context).size.width * 0.9,
-      buttonsBorderRadius: BorderRadius.all(Radius.circular(2)),
-      dismissOnTouchOutside: true,
-      dismissOnBackKeyPress: false,
-      headerAnimationLoop: false,
-      animType: AnimType.SCALE,
-      title: 'Delete Entry?',
-      titleTextStyle: TextStyle(
-        overflow: TextOverflow.ellipsis,
-        color: Colors.green,
-        fontSize: 20,
-        fontWeight: FontWeight.bold,
-      ),
-      desc: 'Are you sure you want to delete this diary entry?',
-      btnOkText: 'Yes',
-      btnOkColor: Color(0xFFE86166),
-      btnCancelColor: Colors.blue,
-      btnCancelText: 'No',
-      btnOkOnPress: () {
-        print('yes');
-      },
-      btnCancelOnPress: () {
-        print('no');
-      },
-      padding: EdgeInsets.all(15),
-      showCloseIcon: false,
-    );
-  }
-
-  //delete successfully
-  AwesomeDialog deleteSuccessDialog(BuildContext context) {
-    return AwesomeDialog(
-      context: context,
-      dialogType: DialogType.SUCCES,
-      borderSide: BorderSide(
-        color: AppColors.secondColor,
-        width: 2,
-      ),
-      width: MediaQuery.of(context).size.width * 0.9,
-      buttonsBorderRadius: BorderRadius.all(Radius.circular(2)),
-      dismissOnTouchOutside: true,
-      dismissOnBackKeyPress: false,
-      headerAnimationLoop: false,
-      animType: AnimType.SCALE,
-      title: 'Deleted Successfully',
-      titleTextStyle: TextStyle(
-        overflow: TextOverflow.ellipsis,
-        color: Colors.green,
-        fontSize: 20,
-        fontWeight: FontWeight.bold,
-      ),
-      onDissmissCallback: (type) {
-        // Navigator.of(context).pop();
-      },
-      padding: EdgeInsets.all(15),
-      showCloseIcon: false,
-      autoHide: Duration(seconds: 3),
-    );
-  }
-
   //deleting entry in list
-  void deleteDiaryEntry(String collectionCaller, String index) {
+  Future<void> deleteDiaryEntry(String collectionCaller, String index) async {
     dbconn.deleteKeyFromCollectionByID(collectionCaller, index);
     print('Deleted diary at index ' + index.toString());
-    deleteSuccessDialog(context).show();
+    showDeletionSuccessDialog(context, 'Diary deleted successfully!');
   }
 
   void editDiaryEntry(BuildContext context, DiaryModel diary) {
-    Navigator.of(context).push(MaterialPageRoute(
-        builder: (context) => EditEntryDiary(selectedDiary: diary)));
+    Navigator.of(context).push(
+      PageTransition(
+        child: EditEntryDiary(selectedDiary: diary,),
+        type: PageTransitionType.rightToLeft,
+      ),
+    );
   }
 }
